+++ src/showTableData.ts
@@ -1,12 +1,7 @@
 import * as vscode from 'vscode';
-import { exit } from 'process';
 import { activeEditor, getALTestRunnerTerminal } from './extension';
 
 export async function showTableData() {
-    if (!activeEditor) {
-        exit;
-    }
-
     let wordAtCursor = getWordAtCursor();
     let recordName = findRecordNameForVariable(wordAtCursor);
     if (recordName !== '') {