+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.5.9
+- Optionally output test results to an editor instead of the output window.
+
+## 0.5.8
+- Respect code warnings
+
 ## 0.5.7
 - Add total code lines, lines hit and percentage to code coverage summary.